Added version to output
