+++ tools/serverpod_cli/bin/serverpod_cli.dart
@@ -14,7 +14,7 @@ import 'internal_tools/generate_pubspecs.dart';
 import 'shared/environment.dart';
 import 'util/command_line_tools.dart';
 import 'util/internal_error.dart';
-import 'util/promt_to_update.dart';
+import 'util/prompt_to_update.dart';
 import 'util/version.dart';
 
 const cmdCreate = 'create';
@@ -62,7 +62,7 @@ Future<void> _main(List<String> args) async {
     return;
   }
 
-  await promtToUpdateIfNeeded();
+  await promptToUpdateIfNeeded();
 
   if (!loadEnvironmentVars()) {
     return;
